smash-batch/requirements.txt
apscheduler==3.10.4
clickhouse-driver==0.2.7
kafka-python==2.0.2